Improve sweep tag handling (#918)
