+++ src/fairseq2/recipes/mt/__init__.py
@@ -10,11 +10,10 @@ from fairseq2.recipes.cli import Cli, RecipeCommandHandler
 from fairseq2.recipes.mt.eval import load_mt_evaluator, mt_eval_presets
 from fairseq2.recipes.mt.train import load_mt_trainer, mt_train_presets
 from fairseq2.recipes.mt.translate import load_text_translator, text_translate_presets
-from fairseq2.recipes.utils.sweep import default_sweep_tagger
 
 
 def _setup_mt_cli(cli: Cli) -> None:
-    default_sweep_tagger.extend_allow_set("source_lang", "target_lang")
+    extra_sweep_keys = {"source_lang", "target_lang"}
 
     group = cli.add_group("mt", help="machine translation recipes")
 
@@ -23,6 +22,7 @@ def _setup_mt_cli(cli: Cli) -> None:
         loader=load_mt_trainer,
         preset_configs=mt_train_presets,
         default_preset="nllb_dense_600m",
+        extra_sweep_keys=extra_sweep_keys,
     )
 
     group.add_command(
@@ -36,6 +36,7 @@ def _setup_mt_cli(cli: Cli) -> None:
         loader=load_mt_evaluator,
         preset_configs=mt_eval_presets,
         default_preset="nllb_dense_600m",
+        extra_sweep_keys=extra_sweep_keys,
     )
 
     group.add_command(
@@ -49,6 +50,7 @@ def _setup_mt_cli(cli: Cli) -> None:
         loader=load_text_translator,
         preset_configs=text_translate_presets,
         default_preset="nllb_dense_600m",
+        extra_sweep_keys=extra_sweep_keys,
     )
 
     group.add_command(
