+++ src/fairseq2/recipes/wav2vec2/__init__.py
@@ -7,7 +7,6 @@
 from __future__ import annotations
 
 from fairseq2.recipes.cli import Cli, RecipeCommandHandler
-from fairseq2.recipes.utils.sweep import default_sweep_tagger
 from fairseq2.recipes.wav2vec2.eval import (
     load_wav2vec2_evaluator,
     wav2vec2_eval_presets,
@@ -19,9 +18,7 @@ from fairseq2.recipes.wav2vec2.train import (
 
 
 def _setup_wav2vec2_cli(cli: Cli) -> None:
-    default_sweep_tagger.extend_allow_set(
-        "max_audio_len", "min_audio_len", "normalize_audio"
-    )
+    extra_sweep_keys = {"max_audio_len", "min_audio_len", "normalize_audio"}
 
     group = cli.add_group("wav2vec2", help="wav2vec 2.0 pretraining recipes")
 
@@ -30,6 +27,7 @@ def _setup_wav2vec2_cli(cli: Cli) -> None:
         loader=load_wav2vec2_trainer,
         preset_configs=wav2vec2_train_presets,
         default_preset="base_960h",
+        extra_sweep_keys=extra_sweep_keys,
     )
 
     group.add_command(
@@ -43,6 +41,7 @@ def _setup_wav2vec2_cli(cli: Cli) -> None:
         loader=load_wav2vec2_evaluator,
         preset_configs=wav2vec2_eval_presets,
         default_preset="base_ls960h",
+        extra_sweep_keys=extra_sweep_keys,
     )
 
     group.add_command(
